Refactor - SnapExpeditedEligibilityParserTest extends AbstractParserTest
