+++ src/test/java/org/codeforamerica/shiba/application/parsers/SnapExpeditedEligibilityParserTest.java
@@ -1,23 +1,13 @@
-package org.codeforamerica.shiba.output.caf;
+package org.codeforamerica.shiba.application.parsers;
 
-import org.codeforamerica.shiba.YamlPropertySourceFactory;
-import org.codeforamerica.shiba.application.parsers.ApplicationDataParser;
-import org.codeforamerica.shiba.application.parsers.ParsingConfiguration;
-import org.codeforamerica.shiba.application.parsers.SnapExpeditedEligibilityParser;
+import org.codeforamerica.shiba.output.caf.JobIncomeInformation;
+import org.codeforamerica.shiba.output.caf.SnapExpeditedEligibilityParameters;
 import org.codeforamerica.shiba.pages.data.ApplicationData;
 import org.codeforamerica.shiba.pages.data.InputData;
 import org.codeforamerica.shiba.pages.data.PageData;
 import org.codeforamerica.shiba.pages.data.PagesData;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.test.context.ActiveProfiles;
 
 import java.util.List;
 import java.util.Map;
@@ -27,40 +17,24 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.NONE;
 
-@SpringBootTest(webEnvironment = NONE)
-@ActiveProfiles("test")
-class SnapExpeditedEligibilityParserTest {
-    @Autowired
-    SnapExpeditedEligibilityParser snapExpeditedEligibilityParser;
-
-    @MockBean
-    ApplicationDataParser<List<JobIncomeInformation>> grossMonthlyIncomeParser;
-
-    @MockBean
-    TotalIncomeCalculator totalIncomeCalculator;
-
-    private final ApplicationData applicationData = new ApplicationData();
-    private final PagesData pagesData = new PagesData();
+class SnapExpeditedEligibilityParserTest extends AbstractParserTest {
+    private SnapExpeditedEligibilityParser snapExpeditedEligibilityParser;
+    private ApplicationData applicationData;
+    private PagesData pagesData;
     private List<JobIncomeInformation> jobIncomeInformation;
-
-    @TestConfiguration
-    @PropertySource(value = "classpath:test-parsing-config.yaml", factory = YamlPropertySourceFactory.class)
-    static class TestPageConfiguration {
-        @SuppressWarnings("ConfigurationProperties")
-        @Bean
-        @ConfigurationProperties(prefix = "test-parsing")
-        public ParsingConfiguration parsingConfiguration() {
-            return new ParsingConfiguration();
-        }
-    }
 
     @BeforeEach
     void setUp() {
+        applicationData = new ApplicationData();
+        pagesData = new PagesData();
         applicationData.setPagesData(pagesData);
+
+        var grossMonthlyIncomeParser = mock(GrossMonthlyIncomeParser.class);
         jobIncomeInformation = List.of(mock(JobIncomeInformation.class));
         when(grossMonthlyIncomeParser.parse(any())).thenReturn(jobIncomeInformation);
+
+        snapExpeditedEligibilityParser = new SnapExpeditedEligibilityParser(parsingConfiguration, grossMonthlyIncomeParser);
     }
 
     @Test
